% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{inmet_extract}
\alias{inmet_extract}
\title{Extract INMET archives}
\usage{
inmet_extract(
  years,
  dest_dir = tools::R_user_dir("rmet", "cache"),
  out_dir = file.path(dest_dir, "csv"),
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{years}{Integer years to extract.}

\item{dest_dir}{Directory containing annual ZIP files.}

\item{out_dir}{Destination directory.}

\item{overwrite}{Overwrite existing files.}

\item{quiet}{Suppress progress messages.}
}
\value{
Invisibly, paths returned by \code{\link[utils:unzip]{utils::unzip()}}.
}
\description{
Extract INMET archives
}
