% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{rmet-package}
\alias{rmet}
\alias{rmet-package}
\title{rmet: Access Historical Weather Data from INMET}
\description{
Tools to discover stations and download, cache, and parse historical hourly
weather data from Brazil's National Institute of Meteorology (INMET).
}
\section{Main functions}{

\itemize{
\item \code{\link[=inmet_get]{inmet_get()}} downloads and reads data in one call.
\item \code{\link[=inmet_download]{inmet_download()}} downloads annual archives.
\item \code{\link[=inmet_read]{inmet_read()}} parses cached archives.
\item \code{\link[=inmet_stations]{inmet_stations()}} searches the official station catalogue.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rodrigosqrt3/rmet}
  \item Report bugs at \url{https://github.com/rodrigosqrt3/rmet/issues}
}

}
\author{
\strong{Maintainer}: Rodrigo Fonseca Villa \email{rodrigo03.villa@gmail.com} (\href{https://orcid.org/0009-0005-2938-2270}{ORCID})

}
\keyword{internal}
